requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
python-frontmatter==1.1.0
python-slugify==8.0.4
PyYAML==5.1
text-unidecode==1.3
toml==0.10.2
Werkzeug==3.0.1
